Move find MPV demo to scripts directory
This removes the pylab import from the find_mpv module.
Matplotlib is no longer required when running the tests.
